fix: use a UUID to track stream
Peer address may not be unique since a single H2 connection can have
multiple streams. Since we do not have any protocol limitation that instructs
a client to open a single stream at any time, it is possible that there
are collisions.

The solution is to generate a UUID on the server side and use that to
track streams.