fix: skip update events to store for writes to node resource
If writes are sent to the database to update the status of a node, it
results in an udpate even pushed to all managers. This results in an
infinite loop, resulting in constant updates of configuration being sent
to Kong Gateway nodes.
After this patch, updates to node resources don't trigger such a
behavior.

This is certainly not the most appropriate solution for the problem at
hand but it is one that works reliably and with minimum effort.

Until we run into problems with this (such as this being used more
elaborately), it is okay.