refactor: flatten foreign relations instead of nested entities
Nested entities work for a few cases but fail for cases when an entity
has multiple parents. Such as Kong's plugin entity.

This refactor removes nesting in favor of flattening foreign relations.
This approach reduces the amount of logic that is required for each
entity, although at some performance cost.