ci: do not continue-on-error
This setting marks the job as successful even if there are failures in
the job.
The intention of adding this flag was to let all jobs to completion
regardless of any siblings in the matrix fail or not.
If such a setting exists in Github Actions, it would be a welcome
change. Until then, this is disabled to avoid false positives.