fix: use protobuf aware JSON marshaller
`valueWrapper` does not implement the proto.Message interface.
When `valueWrapper` is passed to a protobuf-aware JSON marshaller, the
marshaller treats it AND all of the struct members as not having any
protobuf bindings. This results in protobuf unaware JSON marshalling of
`valueWrapper`.

This change breaks down (un)wrapping into two steps:
- protobuf-aware JSON marshalling of the proto.Message
- standard encoding of the valueWrapper

This ensures that the database always has the correct JSON
representation.