fix: enforce additionalProperties in schemas
Commit 2eedc90 highlights a problem: fields can be introduced via
protobuf without corresponding work in JSON schemas.
Upon investigation, this turns out to be a much more serious issue.
Validation for 'additionalProperties: false' is not actually working.
This is because of zero value for 'bool' is 'false' and 'omitempty' is
set for the field.

This patch fixes the problem with the code generator and contains fixes
for all schemas where 'additionalProperties' was being violated.

The fix should also ensure that the same mistake doesn't happen again in
the future.