tests: add cleanup function on GetPersister and cache DB client (#76)
GetPersister returns a 'Persister', associated either to Postgres or
sqlite. Currently, this function has the following problems:

- it doesn't clean the created persisted once done
- it runs queries without context

The above led to the pollution of many idle connections in Postgres,
which ultimately were causing problems in the CI due to the high
a number of tests using this function.

This commit also introduces a clean closure of migrate.Migrate connections.