feat: introduce on-delete cascades for foreign relations (#119)
Up until now, the code errored out if there was a foreign-key relation that
existed on the entity that was being deleted.

With this patch, when the delete operation detects a foreign-key
relation, the related entity is deleted first before deleting the parent
entity. This process is performed recursively - deleting a service
causes the plugins configured on a route related to the service are also
deleted.

This matches up the behavior of DELETE methods in Kong's Admin API with
Koko with one known exception: deleting a ca-certificate resource errors
out when a Service resources references it, koko will instead delete the
Service as well. While not ideal, this is considered acceptable for this patch.