feat: allow field updates for version compatibility (phase 3) (#145)
This feature allows for fields to be updated which meet a condition.
Field updates adds a special condition when adding the config table
updates to the version compatibility layer. A mutually exclusive
condition between Value and ValueFromField is in place as only one or
the other can be defined for a field update.

Note: This is currently used for the zipkin plugin. The integration test
has been updated for zipkin testing as well as verification of
field updates.