feat: export expected hash from the control-plane
The nodes resource contains the hash reported by the DP. A hash is of no
use unless it can be compared against an expected value.
With this patch, the CP stores the hash that it calculates (and also sends
to the DP, which the DP reports back) inside the database.
This hash is available for clients to consume at the
`/v1/expected-config-hash` endpoint.

The caller can compare the response with nodes listed `/v1/nodes`
against the expected config hash to determine if nodes in a cluster are
up-to-date or not.