test: handle version issues with nightly Kong Gateway OSS images
This commit  adds a special case for version parsing issues when using
kong/kong:latest (nightly) image. These Docker images do not return the
proper version of kong as the `kong/meta.lua` file has been modified to
return only the SHA to which commit the Docker image belongs too. In
order to handle testing properly a check for a SHA version number will
occur for data plane nodes connecting and will return a version number
greater that is outside the bounds of a released version.