tests: fix flaky test by using proper util function
For the `TestExpectedConfigHash` test, we are making sure to
run it only when Kong version is >= 2.5.0 by pulling the DP
version from the its Admin API. In order to do so, we need
to wait for the Admin API to be ready to accept requests.

This PR switches the existing util function in use with
another one (`WaitForKongAdminAPI`) doing exactly that.