chore: separate out buf-breaking from lint
`make lint` and `make test` are considered
core and most often used parts of a developer's workflow.
Fetching from remote and running breaking change detection in those
workflow is too slow for a good development experience.
Also, considering that the likelihood a breaking change detected due to
a change in code is likely small and less relevant in a local workflow
as compared to CI.