fix: store pointer to nodes instead of copies
That also means the Ping handler asynchronously modifies a shared resource, and thus needs some form of synchronization.  Currently adding an RWMutex on each node.  It should have a very low contention because each node would only be locked in response of a Ping from the node.

the solutions include:

  1.  a Mutex lock on each node
  2.  RWMutex on each node
  3.  a Mutex lock on each manager
  4.  relegate all modifications to a single goroutine.

Here we use option 2, an RWMutex on each node, as it should have the lowest contention.  It should be write-locked only during the short time it takes to update the struct field.  The space overhead is 24 bytes on each node.  If that proves too high, switching to a simple Mutex would reduce to 8 bytes on each node and rise contention only for the case where a Ping message is received while a new config is sent to the same DP node.

The lowest space overhead would be a Mutex on each manager.  To reduce contention it would be preferable to lock it around the whole update loop for all DP nodes.  Ping messages received during that period could be ignored instead of trying to hold the lock.

Moving updates to a dedicated goroutine amounts to the same serialization as using a lock.

Both in NodeList and mostly everywhere.

From kong/koko#212