perf: lower refresh interval to 5s
The code looks up a change key in the database to detect writes.
This was set to 1s to ease development. The feature is now operating in
a steady state. The value of 5 seconds is chosen to match Kong's default
value of `db_update_frequency`.
The value is lowered conditionally to speed up tests.