fix: relaxing consumer.username validation
The Kong gateway currently has incredibly light validation for the `username`
field on the consumer resource (as of this writing, it allows all unique input).
In order to come closer to its implementation, we're relaxing the validation on
the Koko side of things, to allow for both `@` & `+` characters.

UTF8 support will come at a later time.