fix: make sure ts is used correctly
Right now a check for CreatedAt is done as part of the
lua_validator test suit. This should check whether the
CreatedAt timestamp is less-or-equals to the current time,
but right now this is actually checking the opposite,
making this test to be quite flaky.

This PR makes sure this check is used correctly.