perf: avoid reconciliation of configuration on DP connect
This code predates the event-based refresh of configuration.
Refreshing configuration on a DP-connect was done to simplify
configuration refreshes. With the event-stream now, the configuration is
refreshed dynamically.

The stream is set up before a data-plane is tracked (via AddNode()).
This ensures that the payload will be reconcilied as required and a DP
connect event doesn't require a configuration refresh.

Furthermore, a large number of DP connects could cause a large number of
concurrent configuration reconciliation which is undesired and can even
result in an attack vector.