fix: wrap plugin unmarshal error to not return 500s
When a plugin is validated against the lua VM, an error is
returned if some issue is met. The validator in Koko then attempts
to unmarshal the returned error and to wrap it properly.
In some cases, when the error cannot be properly unmarshalled,
a 500 error code is returned.

This patch makes sure that all validation errors are properly
wrapped as 4xx.