perf: coalesce update events in manager
If there are consistent writes to the database, update events can pile
up on the gRPC stream while configuration is being reconcilied and
pushed down to DP nodes. Working through this pile is not only wasteful
but also results in excessive database queries to reconcile the payload.

This patch changes the design of the Manager by introducing a thread
that "watches" for events.
The thread updating 'updateEventCount' and watching it are separated to
ensure that events are never missed and only one reconiliation takes
place as a time.