Ensure the Payload object is loaded on the first DP
The Manager loads the Payload object via the `m.reconcileKongPayload`
function from the `m.updateEventHandler`, which is executed after
every update event.  But if there hasn't been any such event between
Koko startup and the first DP connection, the Payload object hasn't
had any opportunity to be loaded, thus the first config push would
be empty.

This change adds a call to `m.reconcileKongPayload` when the first
DP node is added to the manager.  It couldn't go inside the
`m.run` function, since it's executed in a coroutine and the config
would get pushed before the payload has been loaded; so we do a
synchronous call.