perf: fetch events only when a node is connected
Previously, update events were fetched once the manager encountered a
node. If the node disconnected, the update events were still fetched,
the configuration was reconciled and cached. This was wasteful if the
manager didn't have any node connected. This can happen frequently if
there are a number of control-plane nodes that are fronted by a
load-balancer.

This commit has a few changes:
- streaming of updates has been pulled out into a new struct `streamer`
- streaming of updates can be disabled or enabled by the manager
- manager no longer runs a dedicated goroutine, goroutine to fetch
  updates is started and stopped as necessary within streamer instead

The overall API of ws.Manager remains unchanged with this patch.