feat: introducing tooling to seed test resources
This introduces tooling that allows the caller (usually integration tests) to
seed test resources in an automated fashion, for a specific resource type or
for all registered resources.

While this tooling was originally created for testing the tag-based listing
integration, it was written in a way to be as versatile as possible, as it's
surely useful in other areas of the codebase as well. I decided to keep
certain parts of the tooling simple given the current use case, and that's
all documented within these changes.

As this tooling is mostly for the sake of integration tests, I did not
specifically write tests against this functionality, but it has been
tested during the development of the tag-based listing integration.