feat: introducing custom JSON resource marshaler on persistence store objects
Currently, there is no way to represent objects within the persistence store
other than the JSON-encoded representation of the object's underlining Protobuf
resource.

This change introduces the ability to override that behavior, in order to store
resources as a custom JSON object, that may diverge from the JSON-encoded
representation of the object's underlining Protobuf resource.