Make wrpc optional
This PR doesn't define any wRPC service apart from the negotiation
service, so it would reject any request as "service unknown"; but the
current Kong only tries the old protocol if the `/v1/wrpc` responds
with 404, even if the negotiation rejects the "config" service.

This feature flag makes this PR safe to deploy by disabling the wRPC
endpoint until the config service is merged.