fix: preventing reference listing with pagination filter
Tag-based listing is currently not supported when used with reference
listing, which is a capability used to filter results based on a
relational object's foreign key.

For example, users can make the following request:
`GET /v1/routes?service_id={service_id}&page.filter=%22tag1%22%20in%20tags`

The above API call would return the routes that match the provided
service ID, and those that have the `tag1` tag.

Prior to this bug fix, such API calls would return no results, as
internally how this works is it makes one query to the persistence
store to grab a list of reference IDs, and then issues DB queries to
fetch each object. The code is improperly applying the tags filter to
the first query, which is incorrect as such records in the persistence
store merely store a relational mapping, not the actual resource data.

As the actual fix for this would require larger changes to the persistence
store, and that we believe tag-based listing won't be used in combination
with this functionality by decK (github.com/Kong/deck), we're preventing
this from happening by throwing an error instead, with an intent to
support this method of listing in the future.