feat: version compatibility for aws-lambda mutually-exclusive fields
According to 3.x changes, this commit removes the mutually-exclusivity
check for the 'aws_region' and 'host' fields. From a version
compatibility perspective, this is not a problem for existing DPs
upgrading to 3.x because they are supposed to run with a good
configuration (which is going to work fine in 3.x as well), but may
be an issue for mixed setup (3.x and 2.x DPs at the same time)
or downgrades.

In order to prevent to impact older DPs, koko will make the call
to drop the 'host' field in the case of both the 'aws_region' and
'host' fields are set.