feat: add changes to 'auth_method' schema field for acme
This commit adds a default value to the 'auth_method' field
for the acme plugin, as well as defines a pool of acceptable
values ('token' and 'kubernetes') for the said field.
Despite this change may break compatibility with older DPs
(which had no limitation for the values this field can be assigned to),
the [docs](https://github.com/fffonion/lua-resty-acme#vault)
were already specifying 'token' and 'kubernetes' as the only
acceptable values.

For this reason, this commit doesn't apply any further version
compatibilty logic to the acme plugin.