remove wrpc-specific code from payload object
Instead of a new typed cache to store prepared wrpc config requests, the
Node object loads the JSON config and performs the wrpc request.

Unfortunately, the existing compressed JSON configs include the
"wrapper" fields, so the wrpc nodes "peel" the config and stores it back
in the config cache. (bypassing the semantic version parser)

It might be better to do this peeling within the version compatibility
feature, at the cost of adding awareness of this requirement to the
semantic version parser.