feat: updating `headers` type on `http-log` plugin schema
In Kong v3.0, the `headers` field on the `http-log` plugin changed from an
array of strings to just a single string. As we want to remain parity with
the latest version of Kong, we're also updating the type within Koko.

While this is a breaking change on the API contract side, we do not want this
to be breaking on the persistence store. As such, we're supporting both ways of
storing the `headers` field, with the expectation of migrating the values in
the persistance store & then removing the backwards compatibile code introduced
in this commit.