feat: introducing `http-log` plugin headers `[]string` -> `string` Postgres migration
This PR complements the changes being done in the below PR:
kong/koko#329

We're redacting the SQLite migration here as currently, the SQLite
persistence store integration is used as an ephemeral persistence store.
While it is technically possible for its data to be saved to disk,
that's not the primary use-case of it right now. As such, we deem it
acceptable to not include the `http-log` `headers` migration for the
SQLite persistence store integration.