feat: introduce new metrics package for instrumentation
'internal/metrics' package provides some abstraction over metrics but it
doesn't provide enough control or abstraction. Specifically:
- no support for subsystem in naming (unless enforced manually
- no support for buckets in histograms

The new package introduces a few functions that wrap Prometheus details.
This is done to ease introduction of new metrics packages, while
providing an implementation that takes full advantages of
Prometheus-specific features.
Support for other metrics backends has been dropped and may be
introduced in the future based on demand.

With the introduction of this new package, 'internal/metrics' is
considered deprecated and will be removed in a following patch.