test: split retry and request timeouts for TestBackoff
With the introduction of larger payloads the data plane does not have
enough time to process and apply the configuration resulting in test
failures due to timeout. This change will increase the request timeout
for the `wait` functions from 5s to 30s in order to give more headroom
to complete applying the payload.