feat: implement internal NodeStatus resource
`node-status` is an internal resource
that keeps track status on a per-node
basis. The ID of node-status resource
and that of node resource are equal.
This resource is not part of the end-user
API-compatibility as it is not exposed
directly to the end-user.
The resource only records enough information
to render a user-facing compatibility status.
Notably, only change IDs and resources are tracked.

Please see the entire patch-set for
additional context.