refactor: late binding of manager on wrpc services
We want to delay the setting of the Manager on wRPC services until a
connection has been authorized.  This means in particular that the
initial configuration (done by cmd/run.go) can't set the manager on the
negotiator object.

The solution is to decouple the registerer object (created at startup)
from the service objects.  This commit splits the Negotiator and
Configurer types into NegotiationRegisterer, negotiationService,
ConfigRegisterer and configService, respectively.

Since the service types do need a reference to the Manager, the Register
function now takes a *Manager parameter, which the ServeHTTP provides,
since it only registers the negotiator after the request has been
authenticated.

From kong/koko#386