feat: 3.0-style prefixed regex paths
For Kong 3.0 any route paths must be prefixed with a `~` to be
considered a regex matching pattern.  It's also explicit and documented
that path matching is done on url-normalized paths.

In Koko the base format will be 3.0-style.  For older versions we strip
the `~` and do a minimal de-normalization (just %% -> %25)