feat: wire up manager to store tracked changes in database
This is the final patch in the series of patch-set for tracking
compatibility related changes.

This patch configures manager to store tracked changes on a per-node
basis. A cache is put in place to reduce the amount of database writes
in case the configuration changes do not impact tracked changes.