refactor: use gcache instead of map for caching configuration
This patch ensures that there exists an upper limit on the number of
cache entries in the cache. Hitting the limit of 10 entries in a
practical use-case is considered unlikely as of this commit since an
operator running more than a few versions of the Kong Gateway with the
same Control-Plane is unlikely.

The cache expiration is meant to release memory in very large scale
deployment.

These are deemed as good starting points and will be tweaked in the
future.