fix: do not allow targets to include subnet mask
Before this commit, Koko considered valid targets being
in the `a.b.c.d/e` format, which actually isn't a valid target
in Kong. This was considered valid because of the regex
used to match an IPv4 format, which was missing the mask part:
for this reason, a target with this format didn't match
the supposed IPv4 regex and it was considered a "domain" target,
undergoing a totally different validation logic than the one
it should have gone through.

This commit is fixing the IPv4 regex to match targets including
a subnet mask, making sure these go through the correct IPv4
validation logic.