fix: no route defaults are applied if 'protocols' is used
From kong/koko#492

Right now, Koko doesn't apply routes' entity schema defaults
when 'protocols' is not empty. The reason for this was
probably an attempt to make schema validation easier: the
`strip_path` default value is `true`, which is not compatible
with `grpc` and `grpcs` protocols; when `protocols` is not empty
defaults are not applied to prevent a incompatibility like
the one about `strip_path` to occur.
This also prevents other defaults to be applied, which is not
desiderable.

This commit removes this restriction by adding a check on the
content of the `protocols` field, setting `strip_path` to `false`
in the case it's not provided as an input.