fix: make service tls_verify a nullable boolean
The admin API only returns the service tls_verify if set to true. If the field is not present or set to false, it is considered as the default value and it is not stored in DB, thus never returned in http responses. This is problematic because that field have 3 meanings:

true: tls verification should happen
false: tls verification should not happen
null/not specified: tls verification decision depends on the DP configuration.

This PR makes that field a nullable boolean, so it is always included in the object definition if set to true or false.

From kong/koko#535