feat: introducing customizable model object cascading functionality
This introduces a new interface that objects can implement in order to
influence its behavior within the persistence store. The primary
use-case for this right now is to override the object cascading
functionality upon deletions.

This change is primarily being introduced for the ongoing consumer
groups effort, though this functionality may come in handy later on.