fix: make unique key check concurrent safe
With multiple concurrent transactions, it is possible that the indexes checks
are done on a old DB state. For object creations, an index that doesn't exist
in that DB state will make koko accept the query and create the new index and
its object in the database. Since this was done concurrently, we can end up
with multiple objects with the same unique index. This is usually handled
properly by postgresql with the primary key unique constraint, but koko is
by-passing this rule with a `on conflict do update ...`.

This change adds a new `Insert` function that can be used in the code and garantee
that an error will be raised if a key already exist at the transaction commit. Now,
the `store.Create` is using it in addition of `createIndexes`.