feat: introducing concept of unmanaged indexes
This allows the ability to bypass the default managed index behavior
that the persistence store provides, in order to explicitly set whether
index(es) should be created and/or removed.

This allows the ability to add & remove foreign relations without
needing to know all foreign relations that exist on the given entity.